Fixed cdf/pdf for inverse_log_uniform
Previous implementation had min/max of the inverse log uniform
distribution specified in the sweep config as the logarithm of
the min and the max you wanted to sample (i.e. in log(x)) space.
The previous implementation was also incorrect.

Now the min/max are given in log(1/x) space and samples from the
pdf/cdf match the true distribution.